Add init.d script to load ssh keys from medium
This fixes #6. At boot, the following actions occur. First, any existing
SSH keys (embedded in the squashfs image) are removed. Next, the boot medium
is checked for a ssh_keys.tar file. If found, this file is extracted into
/etc/ssh. If it is not found, ssh-keygen -A is run to create a new set of
keys, which is stored to ssh_keys.tar on the boot medium for next boot.